Moved custom field serializers to their own module to avoid circular dependency
